+++ packages/types/tsconfig.test.json
@@ -1,0 +1,14 @@
+{
+  "extends": "../../tsconfig.base.json",
+
+  "compilerOptions": {
+    "rootDir": "./",
+
+    "noEmit": true,
+
+    "types": ["node"],
+  },
+
+  "include": ["src"],
+  "exclude": ["**/dist"]
+}
